refactor(mcp): extract resource registrations
